@@ -20,21 +20,21 @@
         }
     ],
     "require": {
-        "php": "^8.0",
+        "php": "^8.4",
         "ext-json": "*",
         "psr/http-client": "^1.0.1",
         "psr/http-message": "^1.0.1 || ^2.0.0"
     },
     "require-dev": {
-        "infection/infection": "^0.26.6",
+        "infection/infection": "^0.31",
         "kriswallsmith/buzz": "^1.2.1",
         "nyholm/psr7": "^1.5.1",
         "php-mock/php-mock-phpunit": "^2.6.0",
-        "phpstan/phpstan": "^1.7.15",
-        "phpunit/phpunit": "^9.5.21",
+        "phpstan/phpstan": "^2.1",
+        "phpunit/phpunit": "^12.0",
         "pimple/pimple": "^3.5.0",
         "rregeer/phpunit-coverage-check": "^0.3.1",
-        "squizlabs/php_codesniffer": "^3.7.1"
+        "squizlabs/php_codesniffer": "^3.11"
     },
     "config": {
         "sort-packages": true,
